+++ src/homematicip/device.py
@@ -2757,3 +2757,7 @@ class MotionDetectorSwitchOutdoor(Device):
 
 class WallMountedKeyPad(Device):
     pass
+
+
+class WateringActuator(Device):
+    pass